fhirqueryapp/requirements.txt
Flask==1.1.2
wtforms
google-api-python-client==1.9.1
google-auth-httplib2==0.0.3
google-auth==1.16.0
google-cloud==0.34.0

gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
